requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.1.31
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
numpy==2.2.4
openai==1.70.0
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.11.1
pydantic_core==2.33.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uvicorn==0.34.0
